fixed issues for stronger_typing
